Rename the occurrence's "services" to "notification_services"
The services.py in the occurrences module included nothing but
notification helper-functions, so it's now renamed to match in that,
since we currently have other services as well.
