+++ occurrences/tests/test_notifications.py
@@ -49,7 +49,7 @@ def test_occurrence_enrolment_notifications_email_only(
     "sms_service_enabled",
     [True, False],
 )
-@patch("occurrences.services.notification_service.send_sms")
+@patch("occurrences.notification_services.notification_service.send_sms")
 def test_occurrence_enrolment_notification_sms_only(
     mock_send_sms,
     sms_service_enabled,
@@ -82,7 +82,7 @@ def test_occurrence_enrolment_notification_sms_only(
 
 
 @pytest.mark.django_db
-@patch("occurrences.services.notification_service.send_sms")
+@patch("occurrences.notification_services.notification_service.send_sms")
 def test_occurrence_enrolment_notification_sms_and_email(
     mock_send_sms,
     snapshot,
@@ -408,7 +408,7 @@ def test_decline_enrolment_notification_email_to_multiple_contact_person(
     [True, False],
 )
 @pytest.mark.django_db
-@patch("occurrences.services.notification_service.send_sms")
+@patch("occurrences.notification_services.notification_service.send_sms")
 def test_mass_approve_enrolment_mutation(
     mock_send_sms,
     full_event,
